% Generated by roxygen2 (4.0.1): do not edit by hand
\name{subsetScenesPlot}
\alias{subsetScenesPlot}
\title{Plot automatically subsetted scenes}
\usage{
subsetScenesPlot(..., bands = NULL, stretch = NULL)
}
\arguments{
\item{...}{Arguments to be passed to \code{\link{subsetScenes}}}

\item{bands}{Numeric. Optional: see \code{details}.}
}
\value{
\code{NULL}, with the side effect of plotting outcome of \code{subsetScenes}
}
\description{
plot scenes selected using \code{subsetScenes}. These can either by single layer or multilayer rasters (or a combination of the two), in which case this function will plot accordingly.
}
\details{
If \code{subsetScenes} returns a multi-layered raster with \code{nlayers > 3}, \code{bands} determines which bands should be plotted (and in which order, according to RGB compositing). If ignored, the first 3 bands will be plotted in that order. If there are 2 layers, the first will be plotted as a single layer.
}
\author{
Ben DeVries
}

