% Generated by roxygen2 (4.0.1): do not edit by hand
\name{tsChips}
\alias{tsChips}
\title{Plot TS Chips}
\usage{
tsChips(x, loc, start = NULL, end = NULL, buff = 17, percNA = 20,
  cols = "PiYG", nbks = 35, nc = 3, nr = 3, ggplot = FALSE)
}
\arguments{
\item{x}{RasterBrick. Image time series brick}

\item{loc}{Location. Can be a vector of length 2 representing the x,y coordinates, or a SpatialPolygons object of \code{nrow = 1} (the first row will be taken if \code{nrow(loc) > 1}), or an extent object (which will be extended if a buffer > 0 is given; see below)}

\item{buff}{Numeric. Number of pixels to buffer the location in all directions. A higher buffer will essentially zoom out.}

\item{start}{Date. OptionaL: earliest date ("yyyy-dd-mm") to display.}

\item{end}{Date. Optional: latest date ("yyyy-dd-mm") to display.}

\item{percNA}{Numeric. Maximum allowable \% NA in the cropped image chips}

\item{cols}{Character. Name of colour map to use (see display.brewer.all()) or a character vector with two or more colour names or hexadecimal values (as strings) between which to interpolate.}

\item{nbks}{Numeric. Number of breaks in the colour map}

\item{nc/nr}{Numeric. Number of columns and rows to plot, respectively. If the number of layers is greater than \code{nc*nr}, a screen prompt will lead to the next series of plots. These cannot exceed 4.}

\item{ggplot}{Logical. Produce a ggplot time series plot object?}
}
\value{
\code{NULL} if \code{ggplot = FALSE} or an object of class \code{ggplot} if \code{ggplot = TRUE}, with the side effect of time series chips being plotted in both cases.
}
\description{
Plot image TS chips based on a location, area buffer, and time buffer
}
\examples{
\dontrun{
library(bfastSpatial)
data(tura)

tsChips(tura, loc = c(820796, 831198))
tsChips(tura, loc = c(820796, 831198), buff = 50) # zoom out
tsChips(tura, loc = c(820796, 831198), buff = 50, percNA = 80) # allow more NA's in field of view
tsChips(tura, loc = c(820796, 831198), start = "2007-01-01", end = "2012-01-01", ggplot = TRUE) # restrict dates and produce pixel time series plot afterwards
}
}
\author{
Ben DeVries
}
\references{
Cohen, W. B., Yang, Z., Kennedy, R. (2010). Detecting trends in forest disturbance and recovery using yearly Landsat time series: 2. TimeSync - Tools for calibration and validation. Remote Sensing of Environment, 114(12), 2911-2924.
}

